django/advanced/requirements.txt
Django==1.7.1
django-image-cropping>=0.9.1